test: Using since in custom template
See https://github.com/appleboy/drone-telegram/issues/89

+++ plugin.go
@@ -59,8 +59,8 @@ type (
 		Number   int
 		Status   string
 		Link     string
-		Started  float64
-		Finished float64
+		Started  int64
+		Finished int64
 		PR       string
 		DeployTo string
 	}
